+++ Doc/Doxygen/Python-API/Source/CDPL/Util/pkg_DG3DCoordinatesGenerator.py
@@ -285,7 +285,7 @@ class DG3DCoordinatesGenerator(Boost.Python.instance):
     ##
     # \brief 
     #
-    DEF_CYCLE_STEP_COUNT_FACTOR = 1
+    DEF_CYCLE_STEP_COUNT_FACTOR = 1.0
 
     ##
     # \brief 
@@ -376,13 +376,13 @@ class DG3DCoordinatesGenerator(Boost.Python.instance):
     # \brief 
     # \param fact 
     #
-    def setCycleStepCountFactor(fact: int) -> None: pass
+    def setCycleStepCountFactor(fact: float) -> None: pass
 
     ##
     # \brief 
     # \return 
     #
-    def getCycleStepCountFactor() -> int: pass
+    def getCycleStepCountFactor() -> float: pass
 
     ##
     # \brief
